+++ src/scripts/setup/options.ts
@@ -10,5 +10,5 @@ export const configTypeOptions: SelectOption<ConfigType | 'scripts'>[] = [
     { label: 'ESLint', value: 'eslint' },
     { label: 'Prettier', value: 'prettier' },
     { label: 'TypeScript', value: 'typescript' },
-    { label: 'Package.json scripts', value: 'scripts' },
+    { label: 'Add package.json scripts', value: 'scripts' },
 ];